Solved c compiling issue and added remove files on error

+++ utils/removeFiles.js
@@ -1,11 +1,11 @@
 const fs = require('fs');
 const path = require('path');
 
-const removeFiles = (filePath, inputPath, fileId, ext) => {
+const removeFiles = (filePath, inputPath, fileId, ext, isCompiled) => {
     const outputPath = path.join(process.cwd(), `outputs/${fileId}.${ext}`);
     fs.unlinkSync(filePath);
     inputPath && fs.unlinkSync(inputPath);
-    ext && fs.unlinkSync(outputPath);
+    isCompiled && ext && fs.unlinkSync(outputPath);
 };
 
 module.exports = removeFiles;
